agents/components/hr/requirements.txt
mcp>=1.25.0
uvicorn>=0.38.0
strands-agents>=1.22.0
boto3>=1.42.0
aws-opentelemetry-distro>=0.10.1
